Refactor web sync engine
